Separate the year and month

Using this function, we can extract the Year and Month information from a seismic catalog for a more detailed analysis. This function works with seismic catalog data where the time_value variable is expressed in the format Year-Month-Day and UTC Time.

SepYM(catalog)

Arguments

catalog

Value

CatalogYM

Details

Catalog YM: The catalog with the year and the month separated

Examples

#data from IGEPN
time_value <- c("2012-01-02 13:02:44.000", "2012-01-02 14:10:00.000", "2012-01-05 01:18:17.000", "2012-01-05 08:43:48.000", "2012-01-05 12:22:12.000", "2012-01-06 02:16:44.000", "2012-01-06 12:20:20.000", "2012-01-08 19:40:29.000", "2012-01-10 18:07:09.000", "2012-01-10 20:54:32.000", "2012-01-12 06:58:55.000", "2012-01-13 20:29:41.000", "2012-01-15 08:22:50.000", "2012-01-15 23:29:17.000", "2012-01-16 11:53:13.000", "2012-01-19 09:47:44.000", "2012-01-19 10:18:02.000", "2012-01-19 10:59:46.000", "2012-01-21 17:11:50.000", "2012-01-22 04:21:59.000", "2012-01-26 19:01:08.000", "2012-01-27 17:28:35.000", "2012-02-01 21:01:28.000", "2012-02-02 08:10:01.000", "2012-02-05 06:13:33.000", "2012-02-05 18:57:42.000", "2012-02-06 17:59:50.000", "2012-02-07 00:59:48.000", "2012-02-07 02:31:22.000", "2012-02-07 04:55:46.000", "2012-02-08 10:54:42.000", "2012-02-09 08:02:20.000", "2012-02-11 02:36:54.000", "2012-02-15 03:46:31.000", "2012-02-18 07:23:27.000", "2012-02-18 07:58:53.000", "2012-02-19 21:16:38.000", "2012-02-20 12:44:47.000", "2012-02-23 22:28:32.000", "2012-02-24 02:56:46.000", "2012-02-26 08:11:47.000", "2012-03-01 03:09:53.000", "2012-03-03 11:10:36.000", "2012-03-05 21:58:26.000", "2012-03-06 18:14:11.000")
depth_value <- c(10, 12, 5, 10, 1.17, 10, 12.83, 31.16, 4.85, 5.15, 6.34, 16.77, 12, 22, 8.9, 12, 30.25, 12, 10, 5.6, 12, 5.8, 10, 19.55, 3.71, 12, 12, 6, 10, 12, 23.21, 25.76, 10, 12, 18.07, 12, 35.36, 12, 12, 136.2, 12, 12, 13.04, 12, 65.42)
magnitude_value_P <- c(3.8, 3.8, 3.8, 3.9, 3.5, 3.8, 3.9, 3.7, 5, 3.7, 4.4, 3.8, 3.6, 4.8, 4.2, 3.6, 5.5, 4.1, 4.2, 3.7, 4, 3.7, 3.7, 3.5, 3.8, 4, 4.3, 4.2, 3.6, 3.8, 5.9, 4, 4.4, 3.7, 4.3, 4.1, 3.8, 3.6, 4.2, 4.9, 3.6, 3.7, 3.8, 3.7, 4.6)
catalog <- data.frame(time_value = time_value, depth_value = depth_value, magnitude_value_P = magnitude_value_P)
SepYM(catalog)
#>    time_value depth_value magnitude_value_P year month
#> 1  2012-01-02       10.00               3.8 2012    01
#> 2  2012-01-02       12.00               3.8 2012    01
#> 3  2012-01-05        5.00               3.8 2012    01
#> 4  2012-01-05       10.00               3.9 2012    01
#> 5  2012-01-05        1.17               3.5 2012    01
#> 6  2012-01-06       10.00               3.8 2012    01
#> 7  2012-01-06       12.83               3.9 2012    01
#> 8  2012-01-08       31.16               3.7 2012    01
#> 9  2012-01-10        4.85               5.0 2012    01
#> 10 2012-01-10        5.15               3.7 2012    01
#> 11 2012-01-12        6.34               4.4 2012    01
#> 12 2012-01-13       16.77               3.8 2012    01
#> 13 2012-01-15       12.00               3.6 2012    01
#> 14 2012-01-15       22.00               4.8 2012    01
#> 15 2012-01-16        8.90               4.2 2012    01
#> 16 2012-01-19       12.00               3.6 2012    01
#> 17 2012-01-19       30.25               5.5 2012    01
#> 18 2012-01-19       12.00               4.1 2012    01
#> 19 2012-01-21       10.00               4.2 2012    01
#> 20 2012-01-22        5.60               3.7 2012    01
#> 21 2012-01-26       12.00               4.0 2012    01
#> 22 2012-01-27        5.80               3.7 2012    01
#> 23 2012-02-01       10.00               3.7 2012    02
#> 24 2012-02-02       19.55               3.5 2012    02
#> 25 2012-02-05        3.71               3.8 2012    02
#> 26 2012-02-05       12.00               4.0 2012    02
#> 27 2012-02-06       12.00               4.3 2012    02
#> 28 2012-02-07        6.00               4.2 2012    02
#> 29 2012-02-07       10.00               3.6 2012    02
#> 30 2012-02-07       12.00               3.8 2012    02
#> 31 2012-02-08       23.21               5.9 2012    02
#> 32 2012-02-09       25.76               4.0 2012    02
#> 33 2012-02-11       10.00               4.4 2012    02
#> 34 2012-02-15       12.00               3.7 2012    02
#> 35 2012-02-18       18.07               4.3 2012    02
#> 36 2012-02-18       12.00               4.1 2012    02
#> 37 2012-02-19       35.36               3.8 2012    02
#> 38 2012-02-20       12.00               3.6 2012    02
#> 39 2012-02-23       12.00               4.2 2012    02
#> 40 2012-02-24      136.20               4.9 2012    02
#> 41 2012-02-26       12.00               3.6 2012    02
#> 42 2012-03-01       12.00               3.7 2012    03
#> 43 2012-03-03       13.04               3.8 2012    03
#> 44 2012-03-05       12.00               3.7 2012    03
#> 45 2012-03-06       65.42               4.6 2012    03